Initial commit for kCARTA-Fortran kCARTA = kCompressed Atmospheric Radiative Transfer Algorithm
kCARTA is a fast, accurate and easy-to-use radiative transfer code.
The package was designed to be the Reference ClearSky Forward Model
for NASA's AIRS instrument, which is a hyperspectral infrared nadir
sounder. The optical depths and radiances in the 605-2830 cm-1 spectral
range (at 0.0025 cm-1 resolution) have been extensively validated during
numerous AIRS campaigns.